StringFormatter: '%d' and '%f' support for the __int__ and __float__ protocol respectively.
The implementation is more convulted than it should be, because we have PyString
implementing __float__ and __int__ at the "java level" but not at the "python
level". For string formatting, only "python level" __float__ and __int__ must be
supported.

Also, considering that __int__ can return a PyLong, this case needs special
care. Basically formatInteger now can call formatLong if a PyLong is found as
the result of calling __int__. Then, as formatLong can also be called from
formatInteger, __hex__, __oct__ and __str__ conversions were moved inside
formatLong.

Finally, test_format_jy was changed to stop checking that we don't support big
floats on '%d' (CPython doesn't, but that seems a limitation of the specific
implementation and I can't imagine a program that could break on Jython because we
*support* it).