Raises NotImplementedError if abstract method from Java is not implemented
Abstract methods of an inherited class or interface from Java now raise
NotImplementedError, instead of returning None (in Java, null) or some
"zero", if they are not implemented in the extending Python class.

This is a (minor) backwards breaking change similar to other recent
changes to remove silent errors in Java integration.

Unlike ABCs in Python - or partial extensions of abstract classes in
Java - we allow such classes to be instantiated. We may wish to change
this in Jython 3.x, but it seems too radical of a change for the moment.

Example: here is the decompiled bytecode for a class implementing
Callable with a method body for the call abstract method like so:

    public Object call() throws Exception {
        final PyObject findPython = ProxyMaker.findPython((PyProxy)this, "call");
        if (findPython != null) {
            final PyObject pyObject = findPython;
            try {
                return Py.tojava(
                    pyObject._jcallexc((Object[])Py.EmptyObjects),
                    (Class)Class.forName("java.lang.Object"));
            }
            catch (Exception ex) {
                throw ex;
            }
            catch (Throwable t) {
                pyObject._jthrow(t);
                return null;
            }
        }
        throw (Throwable)Py.NotImplementedError("");
    }

Previously the last line of this method would have returned null, or
some "zero", per the return type, if it could not look up a method in
the extending Python class.

Fixes http://bugs.jython.org/issue1561