Initial support for ensurepip module
Copies over ensurepip backport from CPython 2.7.9. The main
modification is that Jython needs modified wheels, with the wheels in
wheels in Lib/ensurepip/_bundle built as follows:

* setuptools from setuptools-11.3.1 release
* pip from https://github.com/jythontools/pip

To build wheels, build with ant using a clean release:

1. jython27 ez_setup.py
2. in checkout of https://github.com/jythontools/pip, jython27 setup.py install
3. in checkout of https://github.com/jythontools/wheel, jython27 setup.py install
4. copy output of jpip wheel of pip, setuptools into _bundle

Rebuild Jython and validate with jython27 -m ensurepip. This should be
automated!

This new support needs to be further tested (and likely debugged) on
Windows. Note that some of the tests in test_ensurepip are currently
failing.