Triage of regrtest.py expected skips and failures.
This is a spin-off from work on #2393, carefully distinguishing test
skips (the _expectations variable) from _failures, according to
experience with them on Windows 7x64 and Java 7 (1.7.0_60). YMMV. The
intention is that failures be tracked as issues. Skips *may* be issues
too. Also, that failures affecting a small proportion of a module
be replaced by test_support.skipIf() decorators at module level.