Re-work Py.getJarFileNameFromURL to fix #2410.
This version uses java.net methods instead of ad hoc string manipulation
because these are likely to be correct for both files and network paths.
As a side effect, the file separator is now correct for the platform, so
compensating transformations of the result are removed, and the test
adjusted to expect this.